fix(visibility): address review on #529 (at-least-once reclaim, dedup helper)
Triaged Gemini + Copilot + CodeRabbit + Qodo + the SonarCloud gate.

Real bug (Copilot x2 + Qodo, "stuck claimed events"): the EventLog contract
promises at-least-once re-delivery if a worker crashes between Claim and Ack,
but claimed rows stuck at processed=2 forever. Added a claim lease:
- event_queue gains claimed_at_ns; Claim stamps it and now offers both
  unclaimed rows (processed=0) and rows whose claim is older than claimLeaseNs
  (processed=2, stale), so a crashed worker's events are re-delivered.
- Nack clears claimed_at_ns. New integration test backdates a claim to prove
  the stale-claim reclaim.

SonarCloud gate (new_duplicated_lines_density 3.4% > 3%): the duplicated block
was withDeadlockRetry/isDeadlockErr, copied from the detection store. Extracted
them to server/sqlhelpers (WithDeadlockRetry/IsDeadlockErr) and switched BOTH
the visibility eventlog store and the detection store to the shared helper, so
there is one copy, not two (the reuse rule). Their unit tests moved to
sqlhelpers; detection keeps TestDeduplicateStrings.

Also: wrapped the 145-char INSERT line into an insertPrefix const (Qodo). arch-go
gains sqlhelpers to the visibility.internal allow-list.

Skipped (with reason): json.Marshal of the RawMessage payload (parity with
detection; nil-safe), TINYINT(1) display-width (holds 0/1/2, never scanned to
bool), caller-stamped ingested_at_ns (by design: the ingest handler stamps once
for both stores), CountPending != 1 (bounded table, matches detection), and
INSERT IGNORE error suppression (parity with detection; the truncatable fields
are validated upstream) -- the last is a cross-store hardening follow-up.

Verified: go build, arch-go, golangci-lint (0 issues), and unit + integration
tests for visibility AND detection; coverage 87.4%.
